# Conflicts:
#	build.gradle.kts
#	src/main/kotlin/io/github/starwishsama/comet/Comet.kt
#	src/main/kotlin/io/github/starwishsama/comet/api/command/CommandManager.kt
#	src/main/kotlin/io/github/starwishsama/comet/startup/CometRuntime.kt

+++ src/main/kotlin/io/github/starwishsama/comet/service/compatibility/data/OldGroupConfig.kt
@@ -45,7 +45,7 @@ data class OldGroupConfig(@JsonProperty("group_id") val id: Long) {
      * bilibili 订阅列表
      */
     @JsonProperty("bili_sub")
-    val biliSubscribers: MutableSet<Long> = hashSetOf()
+    val biliSubscribers: MutableSet<Int> = hashSetOf()
 
     /**
      * 是否关闭对此群消息的复读
